Completed homework 16 sprint Заранее спасибо за проверку <3
