+++ app/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
     <uses-permission android:name="android.permission.INTERNET"/>
 
     <application
-        android:name="com.practicum.playlistmaker.presentation.ui.App"
+        android:name="com.practicum.playlistmaker.App"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -16,20 +16,20 @@
         android:theme="@style/Theme.MyApp"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.activities.AudioPlayerActivity"
+            android:name=".player.ui.AudioPlayerActivity"
             android:exported="false"/>
         <activity
-            android:name=".presentation.activities.MediatekaActivity"
+            android:name=".mediateka.ui.activity.MediatekaActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.activities.SearchActivity"
+            android:name=".search.ui.SearchActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustPan"/>
         <activity
-            android:name=".presentation.activities.SettingsActivity"
+            android:name=".settings.ui.activity.SettingsActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.activities.MainActivity"
+            android:name=".main.ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
