--- app/src/main/java/com/practicum/playlistmaker/data/dto/SearchResponse.kt
@@ -1,9 +1,0 @@
-package com.practicum.playlistmaker.data.dto
-
-import com.google.gson.annotations.SerializedName
-import com.practicum.playlistmaker.domain.models.Track
-
-data class SearchResponse(
-    @SerializedName("resultCount") val resultCount: Int,
-    @SerializedName("results") val results: MutableList<Track>
-)
