--- app/src/main/java/com/practicum/playlistmaker/data/network/RetrofitClient.kt
@@ -1,22 +1,0 @@
-package com.practicum.playlistmaker.data.network
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class RetrofitClient {
-    companion object {
-        private var instance: Retrofit? = null
-
-        fun getClient(baseUrl: String): Retrofit {
-            if (instance == null) {
-                instance = Retrofit.Builder()
-                    .baseUrl(baseUrl)
-                    .addConverterFactory(GsonConverterFactory.create())
-                    .build()
-            }
-            return instance!!
-        }
-    }
-}
-
-
