--- app/src/main/java/com/practicum/playlistmaker/data/repositories/MediaPlayerRepositoryImpl.kt
@@ -1,34 +1,0 @@
-package com.practicum.playlistmaker.data.repositories
-
-import android.media.MediaPlayer
-import com.practicum.playlistmaker.domain.api.MediaPlayerRepository
-
-class MediaPlayerRepositoryImpl(private val mediaPlayer: MediaPlayer) : MediaPlayerRepository {
-
-    override fun preparePlayer(previewUrl: String?, onPrepared: () -> Unit, onCompletion: () -> Unit) {
-        mediaPlayer.setDataSource(previewUrl)
-        mediaPlayer.prepareAsync()
-        mediaPlayer.setOnPreparedListener { onPrepared() }
-        mediaPlayer.setOnCompletionListener { onCompletion() }
-    }
-
-    override fun startPlayer() {
-        mediaPlayer.start()
-    }
-
-    override fun pausePlayer() {
-        mediaPlayer.pause()
-    }
-
-    override fun releasePlayer() {
-        mediaPlayer.release()
-    }
-
-    override fun isPlaying(): Boolean {
-        return mediaPlayer.isPlaying
-    }
-
-    override fun getCurrentPosition(): Int {
-        return mediaPlayer.currentPosition
-    }
-}
