--- app/src/main/java/com/practicum/playlistmaker/domain/api/MediaPlayerRepository.kt
@@ -1,10 +1,0 @@
-package com.practicum.playlistmaker.domain.api
-
-interface MediaPlayerRepository {
-    fun preparePlayer(previewUrl: String?, onPrepared: () -> Unit, onCompletion: () -> Unit)
-    fun startPlayer()
-    fun pausePlayer()
-    fun releasePlayer()
-    fun isPlaying(): Boolean
-    fun getCurrentPosition(): Int
-}
