--- app/src/main/java/com/practicum/playlistmaker/domain/impl/MediaPlayerInteractorImpl.kt
@@ -1,31 +1,0 @@
-package com.practicum.playlistmaker.domain.impl
-
-import com.practicum.playlistmaker.domain.api.MediaPlayerInteractor
-import com.practicum.playlistmaker.domain.api.MediaPlayerRepository
-
-class MediaPlayerInteractorImpl(private val repository: MediaPlayerRepository): MediaPlayerInteractor {
-
-    override fun preparePlayer(previewUrl: String?, onPrepared: () -> Unit, onCompletion: () -> Unit) {
-        repository.preparePlayer(previewUrl, onPrepared, onCompletion)
-    }
-
-    override fun startPlayer() {
-        repository.startPlayer()
-    }
-
-    override fun pausePlayer() {
-        repository.pausePlayer()
-    }
-
-    override fun releasePlayer() {
-        repository.releasePlayer()
-    }
-
-    override fun isPlaying(): Boolean {
-        return repository.isPlaying()
-    }
-
-    override fun getCurrentPosition(): Int {
-        return repository.getCurrentPosition()
-    }
-}
