+++ app/src/main/java/com/practicum/playlistmaker/main/ui/MainActivity.kt
@@ -1,0 +1,40 @@
+package com.practicum.playlistmaker.main.ui
+
+import android.content.Intent
+import android.os.Bundle
+import android.widget.Button
+import androidx.appcompat.app.AppCompatActivity
+import com.practicum.playlistmaker.R
+import com.practicum.playlistmaker.mediateka.ui.activity.MediatekaActivity
+import com.practicum.playlistmaker.search.ui.SearchActivity
+import com.practicum.playlistmaker.settings.ui.activity.SettingsActivity
+
+class MainActivity : AppCompatActivity() {
+
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        val searchButton = findViewById<Button>(R.id.search_button)
+        val mediatekaButton = findViewById<Button>(R.id.mediateka_button)
+        val settingsButton = findViewById<Button>(R.id.settings_button)
+
+        searchButton.setOnClickListener {
+            val searchIntent = Intent(this, SearchActivity::class.java)
+            startActivity(searchIntent)
+
+        }
+
+        mediatekaButton.setOnClickListener {
+            val mediatekaIntent = Intent(this, MediatekaActivity::class.java)
+            startActivity(mediatekaIntent)
+        }
+
+        settingsButton.setOnClickListener {
+            val settingsIntent = Intent(this, SettingsActivity::class.java)
+            startActivity(settingsIntent)
+        }
+
+    }
+}
