+++ app/src/main/java/com/practicum/playlistmaker/mediateka/ui/activity/MediatekaActivity.kt
@@ -1,0 +1,12 @@
+package com.practicum.playlistmaker.mediateka.ui.activity
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.practicum.playlistmaker.R
+
+class MediatekaActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_mediateka)
+    }
+}
