+++ app/src/main/java/com/practicum/playlistmaker/player/data/impl/MediaPlayerRepositoryImpl.kt
@@ -1,0 +1,38 @@
+package com.practicum.playlistmaker.player.data.impl
+
+import android.media.MediaPlayer
+import com.practicum.playlistmaker.player.domain.api.MediaPlayerRepository
+
+class MediaPlayerRepositoryImpl(private val mediaPlayer: MediaPlayer) : MediaPlayerRepository {
+
+    override fun preparePlayer(
+        previewUrl: String?,
+        onPrepared: () -> Unit,
+        onCompletion: () -> Unit
+    ) {
+        mediaPlayer.setDataSource(previewUrl)
+        mediaPlayer.prepareAsync()
+        mediaPlayer.setOnPreparedListener { onPrepared() }
+        mediaPlayer.setOnCompletionListener { onCompletion() }
+    }
+
+    override fun startPlayer() {
+        mediaPlayer.start()
+    }
+
+    override fun pausePlayer() {
+        mediaPlayer.pause()
+    }
+
+    override fun releasePlayer() {
+        mediaPlayer.release()
+    }
+
+    override fun isPlaying(): Boolean {
+        return mediaPlayer.isPlaying
+    }
+
+    override fun getCurrentPosition(): Int {
+        return mediaPlayer.currentPosition
+    }
+}
