+++ app/src/main/java/com/practicum/playlistmaker/player/domain/api/MediaPlayerInteractor.kt
@@ -1,0 +1,10 @@
+package com.practicum.playlistmaker.player.domain.api
+
+interface MediaPlayerInteractor {
+    fun preparePlayer(previewUrl: String?, onPrepared: () -> Unit, onCompletion: () -> Unit)
+    fun startPlayer()
+    fun pausePlayer()
+    fun releasePlayer()
+    fun isPlaying(): Boolean
+    fun getCurrentPosition(): Int
+}
