+++ app/src/main/java/com/practicum/playlistmaker/player/domain/models/PlayerState.kt
@@ -1,0 +1,8 @@
+package com.practicum.playlistmaker.player.domain.models
+
+enum class PlayerState {
+    DEFAULT,
+    PREPARED,
+    PLAYING,
+    PAUSED
+}
