--- app/src/main/java/com/practicum/playlistmaker/presentation/activities/MainActivity.kt
@@ -1,35 +1,0 @@
-package com.practicum.playlistmaker.presentation.activities
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.widget.Button
-import com.practicum.playlistmaker.R
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        val searchButton = findViewById<Button>(R.id.search_button)
-        val mediatekaButton = findViewById<Button>(R.id.mediateka_button)
-        val settingsButton = findViewById<Button>(R.id.settings_button)
-
-         searchButton.setOnClickListener {
-            val searchIntent = Intent(this, SearchActivity::class.java)
-             startActivity(searchIntent)
-        }
-
-        mediatekaButton.setOnClickListener {
-            val mediatekaIntent = Intent(this, MediatekaActivity::class.java)
-            startActivity(mediatekaIntent)
-        }
-
-        settingsButton.setOnClickListener {
-            val settingsIntent = Intent(this, SettingsActivity::class.java)
-            startActivity(settingsIntent)
-        }
-
-
-    }
-}
