--- app/src/main/java/com/practicum/playlistmaker/presentation/activities/MediatekaActivity.kt
@@ -1,12 +1,0 @@
-package com.practicum.playlistmaker.presentation.activities
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.practicum.playlistmaker.R
-
-class MediatekaActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_mediateka)
-    }
-}
