+++ app/src/main/java/com/practicum/playlistmaker/search/data/network/RetrofitClient.kt
@@ -1,0 +1,22 @@
+package com.practicum.playlistmaker.search.data.network
+
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+class RetrofitClient {
+    companion object {
+        private var instance: Retrofit? = null
+
+        fun getClient(baseUrl: String): Retrofit {
+            if (instance == null) {
+                instance = Retrofit.Builder()
+                    .baseUrl(baseUrl)
+                    .addConverterFactory(GsonConverterFactory.create())
+                    .build()
+            }
+            return instance!!
+        }
+    }
+}
+
+
