+++ app/src/main/java/com/practicum/playlistmaker/search/domain/api/TrackRepository.kt
@@ -1,0 +1,10 @@
+package com.practicum.playlistmaker.search.domain.api
+
+import com.practicum.playlistmaker.search.domain.models.Track
+
+interface TrackRepository {
+    fun searchTracks(term: String, callback: (Result<MutableList<Track?>>) -> Unit)
+    fun getSearchHistory(): MutableList<Track?>
+    fun addTrackToHistory(track: Track)
+    fun clearSearchHistory()
+}
