+++ app/src/main/java/com/practicum/playlistmaker/search/domain/impl/TrackInteractorImpl.kt
@@ -1,0 +1,24 @@
+package com.practicum.playlistmaker.search.domain.impl
+
+import com.practicum.playlistmaker.search.domain.api.TrackInteractor
+import com.practicum.playlistmaker.search.domain.api.TrackRepository
+import com.practicum.playlistmaker.search.domain.models.Track
+
+class TrackInteractorImpl(private val repository: TrackRepository) : TrackInteractor {
+
+    override fun searchTracks(term: String, callback: (Result<MutableList<Track?>>) -> Unit) {
+        repository.searchTracks(term, callback)
+    }
+
+    override fun getSearchHistory(): MutableList<Track?> {
+        return repository.getSearchHistory()
+    }
+
+    override fun addTrackToHistory(track: Track) {
+        repository.addTrackToHistory(track)
+    }
+
+    override fun clearSearchHistory() {
+        repository.clearSearchHistory()
+    }
+}
