+++ app/src/main/java/com/practicum/playlistmaker/search/domain/models/Track.kt
@@ -1,0 +1,66 @@
+package com.practicum.playlistmaker.search.domain.models
+
+import android.os.Parcel
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+
+
+data class Track(
+    @SerializedName("trackId") val trackId: Long,
+    @SerializedName("trackName") val trackName: String?,
+    @SerializedName("artistName") val artistName: String?,
+    @SerializedName("trackTimeMillis") val trackTimeMillis: Long,
+    @SerializedName("artworkUrl100") val artworkUrl100: String?,
+    @SerializedName("collectionName") val collectionName: String?,
+    @SerializedName("releaseDate") val releaseDate: String?,
+    @SerializedName("primaryGenreName") val primaryGenreName: String?,
+    @SerializedName("country") val country: String?,
+    @SerializedName("previewUrl") val previewUrl: String?
+) : Parcelable {
+
+    val artworkUrl512 = artworkUrl100?.replaceAfterLast('/', "512x512bb.jpg")
+
+
+    constructor(parcel: Parcel) : this(
+        parcel.readLong(),
+        parcel.readString(),
+        parcel.readString(),
+        parcel.readLong(),
+        parcel.readString(),
+        parcel.readString(),
+        parcel.readString(),
+        parcel.readString(),
+        parcel.readString(),
+        parcel.readString()
+    ) {
+    }
+
+    override fun writeToParcel(parcel: Parcel, flags: Int) {
+        parcel.writeLong(trackId)
+        parcel.writeString(trackName)
+        parcel.writeString(artistName)
+        parcel.writeLong(trackTimeMillis)
+        parcel.writeString(artworkUrl100)
+        parcel.writeString(collectionName)
+        parcel.writeString(releaseDate)
+        parcel.writeString(primaryGenreName)
+        parcel.writeString(country)
+        parcel.writeString(previewUrl)
+    }
+
+    override fun describeContents(): Int {
+        return 0
+    }
+
+
+    companion object CREATOR : Parcelable.Creator<Track> {
+        override fun createFromParcel(parcel: Parcel): Track {
+            return Track(parcel)
+        }
+
+        override fun newArray(size: Int): Array<Track?> {
+            return arrayOfNulls(size)
+        }
+    }
+
+}
