+++ app/src/main/java/com/practicum/playlistmaker/search/ui/adapters/TrackViewHolder.kt
@@ -1,0 +1,44 @@
+package com.practicum.playlistmaker.search.ui.adapters
+
+import android.view.View
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.practicum.playlistmaker.R
+import com.practicum.playlistmaker.search.domain.models.Track
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class TrackViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    private val trackName: TextView = itemView.findViewById(R.id.trackName)
+    private val artistName: TextView = itemView.findViewById(R.id.artistName)
+    private val trackTime: TextView = itemView.findViewById(R.id.trackTime)
+    private val artworkUrl100: ImageView = itemView.findViewById(R.id.artworkUrl100)
+
+    companion object {
+        const val ROUNDED_CORNER_RADIUS = 12
+    }
+
+    fun bind(track: Track) {
+
+        trackName.text = track.trackName
+        artistName.text = track.artistName
+        trackTime.text =
+            SimpleDateFormat("mm:ss", Locale.getDefault()).format(track.trackTimeMillis)
+
+
+        if (track.artworkUrl100?.isNotEmpty() == true) {
+            Glide.with(itemView.context)
+                .load(track.artworkUrl100)
+                .placeholder(R.drawable.placeholder)
+                .transform(RoundedCorners(ROUNDED_CORNER_RADIUS))
+                .into(artworkUrl100)
+        } else {
+            artworkUrl100.setImageResource(R.drawable.placeholder)
+        }
+
+
+    }
+}
