+++ app/src/main/java/com/practicum/playlistmaker/settings/domain/SettingsInteractor.kt
@@ -1,0 +1,9 @@
+package com.practicum.playlistmaker.settings.domain
+
+import com.practicum.playlistmaker.settings.domain.model.ThemeSettings
+
+interface SettingsInteractor {
+
+    fun getThemeSettings(): ThemeSettings
+    fun updateThemeSettings(isChecked: Boolean)
+}
