+++ app/src/main/java/com/practicum/playlistmaker/settings/domain/SettingsRepository.kt
@@ -1,0 +1,9 @@
+package com.practicum.playlistmaker.settings.domain
+
+import com.practicum.playlistmaker.settings.domain.model.ThemeSettings
+
+interface SettingsRepository {
+
+    fun getThemeSettings(): ThemeSettings
+    fun updateThemeSettings(settings: ThemeSettings)
+}
