+++ app/src/main/java/com/practicum/playlistmaker/settings/domain/impl/SettingsInteractorImpl.kt
@@ -1,0 +1,17 @@
+package com.practicum.playlistmaker.settings.domain.impl
+
+import com.practicum.playlistmaker.settings.domain.SettingsInteractor
+import com.practicum.playlistmaker.settings.domain.SettingsRepository
+import com.practicum.playlistmaker.settings.domain.model.ThemeSettings
+
+class SettingsInteractorImpl(
+    private val settingsRepository: SettingsRepository
+) : SettingsInteractor {
+    override fun getThemeSettings(): ThemeSettings {
+        return settingsRepository.getThemeSettings()
+    }
+
+    override fun updateThemeSettings(isChecked: Boolean) {
+        return settingsRepository.updateThemeSettings(ThemeSettings(isChecked))
+    }
+}
