+++ app/src/main/java/com/practicum/playlistmaker/settings/domain/model/ThemeSettings.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker.settings.domain.model
+
+data class ThemeSettings(
+    val darkTheme: Boolean
+)
