+++ app/src/main/java/com/practicum/playlistmaker/share/domain/ExternalNavigator.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker.share.domain
+
+interface ExternalNavigator {
+    fun share()
+    fun openLink()
+    fun sendEmail()
+}
