+++ app/src/main/java/com/practicum/playlistmaker/share/domain/SharingInteractor.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker.share.domain
+
+interface SharingInteractor {
+    fun openTerms()
+    fun shareApp()
+    fun contactSupport()
+}
