+++ app/src/main/java/com/practicum/playlistmaker/share/domain/impl/SharingInteractorImpl.kt
@@ -1,0 +1,20 @@
+package com.practicum.playlistmaker.share.domain.impl
+
+import com.practicum.playlistmaker.share.domain.ExternalNavigator
+import com.practicum.playlistmaker.share.domain.SharingInteractor
+
+class SharingInteractorImpl(
+    private val externalNavigator: ExternalNavigator,
+) : SharingInteractor {
+    override fun openTerms() {
+        externalNavigator.openLink()
+    }
+
+    override fun shareApp() {
+        externalNavigator.share()
+    }
+
+    override fun contactSupport() {
+        externalNavigator.sendEmail()
+    }
+}
