+++ app/src/main/java/com/practicum/playlistmaker/share/domain/model/EmailData.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker.share.domain.model
+
+data class EmailData(
+    val mailTo: String = "mailto:"
+)
