+++ app/src/main/res/layout/activity_mediateka.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.activities.MediatekaActivity">
+    tools:context=".mediateka.ui.activity.MediatekaActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
